LayoutTests/imported/w3c/web-platform-tests/webrtc/legacy/munge-dont-expected.txt


PASS RTCSessionDescription.type is read-only
PASS RTCSessionDescription.sdp is read-only
PASS RTCIceCandidate.candidate is read-only
PASS Rejects SDP munging between createOffer and setLocalDescription
PASS Rejects SDP munging between createAnswer and setLocalDescription
